First, let's be less clever
Skipping the use of handle_out back in the day was a design mistake, as
was stripping nonpresenter data recursively when I could just conver the
top-level slide to a map when needed to bring in presenter data. To be
fair, I was much dumber when I thought I was clever.